Service Chiefs (primarily) and some refactoring of common methods
