+++ test/Test-Automation/test-output/RegressionSuite/Service Chiefs.xml
@@ -1,0 +1,52 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitXMLReporter -->
+<testsuite ignored="0" hostname="DESKTOP-OOHKJP3" failures="1" tests="15" name="Service Chiefs" time="32.426" errors="0" timestamp="2023-09-14T20:10:50 CDT">
+  <testcase classname="Execution.serviceChiefs" name="testForCertPage" time="1.525"/>
+  <testcase classname="Execution.serviceChiefs" name="clickSetupWizard" time="2.91"/>
+  <testcase classname="Execution.serviceChiefs" name="clickExecLeadershipTeam" time="2.563"/>
+  <testcase classname="Execution.serviceChiefs" name="clickCreateELT" time="1.059"/>
+  <testcase classname="Execution.serviceChiefs" name="enterNameOfService" time="3.849"/>
+  <testcase classname="Execution.serviceChiefs" name="enterPositionTitle" time="2.074"/>
+  <testcase classname="Execution.serviceChiefs" name="inputEmployeeName" time="2.521"/>
+  <testcase classname="Execution.serviceChiefs" name="saveSCGroup" time="0.362"/>
+  <testcase classname="Execution.serviceChiefs" name="createNexusDriver1" time="5.812"/>
+  <testcase classname="Execution.serviceChiefs" name="scrollDownNexus" time="2.013"/>
+  <testcase classname="Execution.serviceChiefs" name="syncServices" time="1.792"/>
+  <testcase classname="Execution.serviceChiefs" name="navigateNexusBack" time="0.262"/>
+  <testcase classname="Execution.serviceChiefs" name="clickServiceChiefs" time="1.415"/>
+  <testcase classname="Execution.serviceChiefs" name="scrollUpNexus" time="0.523"/>
+  <testcase classname="Execution.serviceChiefs" name="openSCGroup" time="1.036">
+    <failure type="org.openqa.selenium.NoSuchElementException" message="no such element: Unable to locate element: {&amp;quot;method&amp;quot;:&amp;quot;xpath&amp;quot;,&amp;quot;selector&amp;quot;:&amp;quot;//*[contains(text(),&amp;apos;000 Automated Test Services&amp;apos;)]&amp;quot;}
+  (Session info: chrome=116.0.5845.188)
+For documentation on this error, please visit: https://www.seleniumhq.org/exceptions/no_such_element.html
+Build info: version: &amp;apos;3.141.59&amp;apos;, revision: &amp;apos;e82be7d358&amp;apos;, time: &amp;apos;2018-11-14T08:25:48&amp;apos;
+System info: host: &amp;apos;DESKTOP-OOHKJP3&amp;apos;, ip: &amp;apos;192.168.1.223&amp;apos;, os.name: &amp;apos;Windows 10&amp;apos;, os.arch: &amp;apos;amd64&amp;apos;, os.version: &amp;apos;10.0&amp;apos;, java.version: &amp;apos;15.0.2&amp;apos;
+Driver info: org.openqa.selenium.chrome.ChromeDriver
+Capabilities {acceptInsecureCerts: false, browserName: chrome, browserVersion: 116.0.5845.188, chrome: {chromedriverVersion: 116.0.5845.96 (1a3918166880..., userDataDir: C:\Users\MAXRIC~1\AppData\L...}, fedcm:accounts: true, goog:chromeOptions: {debuggerAddress: localhost:63184}, javascriptEnabled: true, networkConnectionEnabled: false, pageLoadStrategy: normal, platform: WINDOWS, platformName: WINDOWS, proxy: Proxy(), setWindowRect: true, strictFileInteractability: false, timeouts: {implicit: 0, pageLoad: 300000, script: 30000}, unhandledPromptBehavior: dismiss and notify, webauthn:extension:credBlob: true, webauthn:extension:largeBlob: true, webauthn:extension:minPinLength: true, webauthn:extension:prf: true, webauthn:virtualAuthenticators: true}
+Session ID: 0092c0708900fb7e4a4686172e6e7914
+*** Element info: {Using=xpath, value=//*[contains(text(),&amp;apos;000 Automated Test Services&amp;apos;)]}">
+      <![CDATA[org.openqa.selenium.NoSuchElementException: no such element: Unable to locate element: {"method":"xpath","selector":"//*[contains(text(),'000 Automated Test Services')]"}
+  (Session info: chrome=116.0.5845.188)
+For documentation on this error, please visit: https://www.seleniumhq.org/exceptions/no_such_element.html
+Build info: version: '3.141.59', revision: 'e82be7d358', time: '2018-11-14T08:25:48'
+System info: host: 'DESKTOP-OOHKJP3', ip: '192.168.1.223', os.name: 'Windows 10', os.arch: 'amd64', os.version: '10.0', java.version: '15.0.2'
+Driver info: org.openqa.selenium.chrome.ChromeDriver
+Capabilities {acceptInsecureCerts: false, browserName: chrome, browserVersion: 116.0.5845.188, chrome: {chromedriverVersion: 116.0.5845.96 (1a3918166880..., userDataDir: C:\Users\MAXRIC~1\AppData\L...}, fedcm:accounts: true, goog:chromeOptions: {debuggerAddress: localhost:63184}, javascriptEnabled: true, networkConnectionEnabled: false, pageLoadStrategy: normal, platform: WINDOWS, platformName: WINDOWS, proxy: Proxy(), setWindowRect: true, strictFileInteractability: false, timeouts: {implicit: 0, pageLoad: 300000, script: 30000}, unhandledPromptBehavior: dismiss and notify, webauthn:extension:credBlob: true, webauthn:extension:largeBlob: true, webauthn:extension:minPinLength: true, webauthn:extension:prf: true, webauthn:virtualAuthenticators: true}
+Session ID: 0092c0708900fb7e4a4686172e6e7914
+*** Element info: {Using=xpath, value=//*[contains(text(),'000 Automated Test Services')]}
+at org.openqa.selenium.remote.http.W3CHttpResponseCodec.createException(W3CHttpResponseCodec.java:187)
+at org.openqa.selenium.remote.http.W3CHttpResponseCodec.decode(W3CHttpResponseCodec.java:122)
+at org.openqa.selenium.remote.http.W3CHttpResponseCodec.decode(W3CHttpResponseCodec.java:49)
+at org.openqa.selenium.remote.HttpCommandExecutor.execute(HttpCommandExecutor.java:158)
+at org.openqa.selenium.remote.service.DriverCommandExecutor.execute(DriverCommandExecutor.java:83)
+at org.openqa.selenium.remote.RemoteWebDriver.execute(RemoteWebDriver.java:552)
+at org.openqa.selenium.remote.RemoteWebDriver.findElement(RemoteWebDriver.java:323)
+at org.openqa.selenium.remote.RemoteWebDriver.findElementByXPath(RemoteWebDriver.java:428)
+at org.openqa.selenium.By$ByXPath.findElement(By.java:353)
+at org.openqa.selenium.remote.RemoteWebDriver.findElement(RemoteWebDriver.java:315)
+at Execution.serviceChiefs.openSCGroup(serviceChiefs.java:399)
+at java.base/java.util.ArrayList.forEach(ArrayList.java:1511)
+... Removed 32 stack frames]]>
+    </failure>
+  </testcase> <!-- openSCGroup -->
+</testsuite> <!-- Service Chiefs -->
