+++ test/Test-Automation/test-output/old/RegressionSuite/User Access Groups Part 2.properties
@@ -1,1 +1,1 @@
-[SuiteResult context=User Access Groups][SuiteResult context=User Access Groups Part 2]
+[SuiteResult context=User Access Groups TESTING][SuiteResult context=User Access Groups Part 2]
