+++ test/Test-Automation/test-output/old/RegressionSuite/toc.html
@@ -10,7 +10,7 @@
 <tr valign='top'>
 <td>1 test</td>
 <td><a target='mainFrame' href='classes.html'>1 class</a></td>
-<td>37 methods:<br/>
+<td>15 methods:<br/>
 &nbsp;&nbsp;<a target='mainFrame' href='methods.html'>chronological</a><br/>
 &nbsp;&nbsp;<a target='mainFrame' href='methods-alphabetical.html'>alphabetical</a><br/>
 &nbsp;&nbsp;<a target='mainFrame' href='methods-not-run.html'>not run (0)</a></td>
@@ -20,10 +20,10 @@
 <td><a target='mainFrame' href='reporter-output.html'>reporter output</a></td>
 <td><a target='mainFrame' href='testng.xml.html'>testng.xml</a></td>
 </tr></table>
-<table width='100%' class='test-passed'>
+<table width='100%' class='test-failed'>
 <tr><td>
-<table style='width: 100%'><tr><td valign='top'>User Access Groups TESTING (37/0/0)</td><td valign='top' align='right'>
-  <a href='User Access Groups TESTING.html' target='mainFrame'>Results</a>
+<table style='width: 100%'><tr><td valign='top'>Service Chiefs (14/1/0)</td><td valign='top' align='right'>
+  <a href='Service Chiefs.html' target='mainFrame'>Results</a>
 </td></tr></table>
 </td></tr><p/>
 </table>
